Refactoring: decoupling QueryExecutor from Database (to use DataSource instead) and extracting query execution logic from ExtractExecutor to ExtractQueryExecutor.
